* Update to r1096333 of the svn trunk * Adjust for the changed update and switch API * add new checkbox fo the update dialog to have the local mods to svn:externals applied
The default (because faster and with less problems for file externals) is to *not* apply local mods to svn:externals. So to have those applied, the "update to revision" dialog has to be used.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21223 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlg.cpp
@@ -495,7 +495,7 @@ BOOL CLogDlg::OnInitDialog()
     m_ChangedFileListCtrl.GetWindowRect(&rcChgMsg);
     ScreenToClient(&rcChgMsg);
 
-    if (yPos1 && yPos1 < rcDlg.bottom - 185)
+    if (yPos1 && ((LONG)yPos1 < rcDlg.bottom - 185))
     {
         RECT rectSplitter;
         m_wndSplitter1.GetWindowRect(&rectSplitter);
@@ -508,7 +508,7 @@ BOOL CLogDlg::OnInitDialog()
             DoSizeV1(delta);
         }
     }
-    if (yPos2 && yPos2 < rcDlg.bottom - 153)
+    if (yPos2 && ((LONG)yPos2 < rcDlg.bottom - 153))
     {
         RECT rectSplitter;
         m_wndSplitter2.GetWindowRect(&rectSplitter);
